Fix incorrect checking of closing paren in primary
Closing paren was always checked for at the end of the line. This will
not work correctly in cases such as: (5)+2.

Changes were also made to reverse_search which is now used
inside primary to find the closing paren. reverse_search is now more
general with the option to turn nesting, and previous operator
checking off.