Correct reverse_search's doccomment
The incorrect doccomment says that either the prev_op_ignore
vector is honored OR preceding operators are ignored.

In fact, both is done: a found letter
is not ignored if it doesn't follows an operator. It's also not ignored if
the operator it follows occurs in the
prev_op_ignore list.