Fix issue #1
This issue caused inputs where the factorial was combined with
other operations to fail. For example, "4! + 3!" would fail. This
happened because `reverse_search` would ignore
the `+` sign because it occurred after an operator (i.e., `!`). The fix
was to add `!` to the list of operators that would be allowed
to precede the operator that occurs in `to_find`.